Ruleset for file requests v0.1 - the basics
Existing files should be delivered by apache without any processing. If
an asset is missing it should trigger the creation process and pass on
any possible parameters from the request so that we can create
and optimize the asset for instant delivery on next roundtrip.

For next version, cache and expiry headers for asset delivery.